added wrapper class MessagePlus to be used with MessageManager and ADNDatabase
So far, this class supports using a custom display date and display location. The point of
having these methods is to improve performance in situations where you will need to retrieve
annotations associated with these values repeatedly. For example, when scrolling a ListView,
it is not especially performant to have to repeatedly retrieve a geolocation annotation, use a
Geocoder to get associated addresses, and then build a String to associate with that location.
Thus, you can use hasSetDisplayLocation to determine if such an algorithm has to be used.